Add Kazuyoshi Kato as committer
Kazu has been actively contributing to containerd for awhile now. As
contributor, he has taken on solving complex issues and contributing
significant features to containerd. He has been one of the most active
in the community helping users with their issues and triaging bugs.
Quality is very important to containerd and we have adopted a strict
review process to maintain it. Kazu has been very helpful in
providing frequent and in-depth code reviews to help us maintain the
quality of containerd and our associated projects. All this qualifies
Kazu to move to the committer role, which is primarily differentiated
by the responsibility to maintain the overall quality and direction
of the containerd project.

Signed-off-by: Derek McGowan <derek@mcg.dev>