Add Samuel Karp as committer
Sam is a long time contributor and user of containerd. He has helped out
in many roles including reviewer, security advisor, as well as helping
out with design feedback and getting directly involved in the security
release process. He has shown a strong commitment to maintaining the
overall project quality and understanding of the project scope.

Signed-off-by: Derek McGowan <derek@mcg.dev>